requirements.txt
argh==0.26.2
asn1crypto==0.23.0
awscli==1.16.210
backports.ssl-match-hostname==3.7.0.1
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.218
botocore==1.12.200
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.7
dictop==0.1.4
Django==2.1.1
django-admin-honeypot==1.1.0
django-auth==0.1.9
django-bootstrap==0.2.4
django-braces==1.13.0
django-cleanup==4.0.0
django-click==2.1.0
django-crispy-forms==1.7.2
django-csvimport==2.12
django-dynamodb-sessions==0.6
django-environ==0.4.5
django-extended-choices==1.3.3
django-extensions==2.2.1
django-fields==0.3.0
django-filter==2.1.0
django-guardian==2.0.0
django-ipware==2.1.0
django-ipyfield==0.1.6
django-jet==1.0.8
django-jinja2==0.1
django-livereload-server==0.3.2
django-livesync==0.4
django-macaddress==1.6.0
django-model-utils==3.2.0
django-multiselectfield==0.1.8
django-pipeline==1.6.14
django-query==1.0.3
django-ranged-response==0.2.0
django-request==1.5.4
django-rest-framework==0.1.0
django-reversion==3.0.4
django-select-multiple-field==0.4.2
django-selectable==1.2.1
django-simple-captcha==0.5.12
django-split-settings==0.3.0
django-storages==1.7.1
django-submodel==0.1
django-test-without-migrations==0.6
django-utils==0.0.2
django-warrant==0.1.0
djangorestframework==3.7.7
djangorestframework-camel-case==1.0.3
docutils==0.14
docker-py==1.10.6
docker-pycreds==0.4.0
ecdsa==0.13.2
envs==1.3
future==0.17.1
getmac==0.8.1
gunicorn==19.9.0
idna==2.6
IPy==1.0
jmespath==0.9.4
magic-import==0.1.3
mock==3.0.5
netaddr==0.7.19
numpy==1.17.4
pathtools==0.1.2
Pillow==6.1.0
pyasn1==0.4.6
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.3.1
PyJWT==1.6.4
pyOpenSSL==19.0.0
pypi==2.1
python-dateutil==2.6.1
python-jose-cryptodome==1.3.2
python-pkg==0.0.2
pytz==2019.1
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.3
South==1.0.2
sqlparse==0.3.0
tornado==4.5.1
urllib3==1.22
uWSGI==2.0.18
warrant==0.6.1
watchdog==0.9.0
websocket-client==0.56.0

